improve jackson

+++ src/main/java/com/scut/cs/service/ProjectsServiceImpl.java
@@ -1,12 +1,16 @@
 package com.scut.cs.service;
 
 import com.scut.cs.domain.Project;
+import com.scut.cs.domain.Student;
 import com.scut.cs.domain.dao.ProjectRepository;
+import com.scut.cs.domain.dao.StudentRepository;
+import com.scut.cs.web.request.AddStudents;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.annotation.security.RolesAllowed;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
@@ -19,11 +23,15 @@ public class ProjectsServiceImpl implements ProjectsService {
 
     @Autowired
     private ProjectRepository projectRepository;
+    @Autowired
+    private StudentRepository studentRepository;
 
     @PreAuthorize("hasRole('ROLE_INNER') or hasRole('ROLE_ADMIN')")
     @Override
     public List<Project> getAllProjects() {
         return projectRepository.findAll();//TODO 分页和排序
+
+//        return null;
     }
 
     @PreAuthorize("hasRole('ROLE_OUTER') or hasRole('ROLE_OUTER_SPEC')")
@@ -39,9 +47,36 @@ public class ProjectsServiceImpl implements ProjectsService {
     @Override
     public Project addProject(Project project) {
         if (project != null) {
-            Project projectOle = projectRepository.withNameAndRankAndDateQuery(project.getProjectName(), project.getRank(), project.getProjectDate());
-            if ( null == projectOle) {
+            Project projectOld = projectRepository.withNameAndRankAndDateQuery(project.getProjectName(), project.getRank(), project.getProjectDate());
+            if ( null == projectOld) {
                 return projectRepository.save(project);
+            }
+        }
+        return null;
+    }
+
+    @PreAuthorize("hasRole('ROLE_OUTER')")
+    @Override
+    public Project addStudents(AddStudents addStudents) {
+        if (null != addStudents && null != addStudents.getProjectId() && 0 != addStudents.getProjectId()) {
+            Long projectId=addStudents.getProjectId();
+            Project project=projectRepository.findOne(projectId);
+            if (null!= project) {
+                if (project.getId()%2==1) {
+                    project.getStudentList().addAll(addStudents.getStudentList());
+                    return projectRepository.save(project);
+                } else {
+                    List<Student> studentList = new ArrayList<>();
+                    for (Student student : addStudents.getStudentList()) {
+                        List<Project> projectList = new ArrayList<>();
+                        projectList.add(project);
+                        student.setProjectList(projectList);
+                        studentList.add(student);
+                    }
+                    studentRepository.save(studentList);
+                    return project;
+                }
+
             }
         }
         return null;
@@ -61,7 +96,7 @@ public class ProjectsServiceImpl implements ProjectsService {
     @PreAuthorize("hasRole('ROLE_INNER') or hasRole('ROLE_ADMIN')")
     @Override
     public int changeProjectState(Long id, String state) {
-        if (id > 0 && state != null && !state.equals("")) {
+        if (null!=id && id > 0 && state != null && !state.equals("")) {
             if (projectRepository.exists(id)) {
                 return projectRepository.setProjectState(id,state);
             }
@@ -72,7 +107,7 @@ public class ProjectsServiceImpl implements ProjectsService {
     @PreAuthorize("hasRole('ROLE_OUTER') or hasRole('ROLE_ADMIN')")
     @Override
     public Long deleteProject(Long id) {
-        if (id>0 && projectRepository.exists(id)) {
+        if (id != null && id>0 && projectRepository.exists(id)) {
             projectRepository.delete(id);
             return id;
         }
@@ -83,11 +118,11 @@ public class ProjectsServiceImpl implements ProjectsService {
     @Transactional(rollbackFor = {IllegalArgumentException.class})
     @Override
     public List<Long> deleteProjectList(List<Long> idList) {
-        if (null != idList && !idList.isEmpty()) {
+        if (null != idList && 0!=idList.size()) {
             List<Long> projectIds = new ArrayList<>();
             for (Long id : idList) {
                 if (!Objects.equals(id, deleteProject(id))) {
-                    throw new IllegalArgumentException("项目id为："+id+" 的项目不存在！！");   //TODO 验证能不能继续正常执行程序
+                    throw new IllegalArgumentException("项目id为："+id+" 的项目不存在！！");   //TODO 可以在controller捕抓异常
                 }
                 projectIds.add(id);
             }
